chore: split conversion into multiple files
zero code changes
