+++ errors.go
@@ -1,0 +1,5 @@
+package safecast
+
+import "errors"
+
+var ErrOutOfRange = errors.New("out of range")
